Views and controllers for rendering department-level pages
Added functionality that renders department-level pages and slugs from
Twig. However, it’s a bit sketchy how it’s coming in from Twig. Per
recommendations, I’m going to create a new schema for Departments and
include a column just for slugs. This will help with controller logic
that generates slugs, and will add objects to array that looks up staff
members associated with that department. This will also get the route
generation out of Twig.